rush update --full 8/26/2021 (#17294)
I was trying to remove some of the unused dependencies in our packages and this invariably will change the lock file. Found that `@types/node` v12 had a patch update in the past 24 hours that caused much noise to my updates. So, creating this PR with just `rush update --full` so that my future PRs can have less noise

Please note that the automated weekly PR that does this was already run yesterday and merged