[KeyVault] - Add getCryptographyClient method to KeyClient (#17529)
## What

- Adds the `KeyClient#getCryptographyClient` convenience method 
- Updates remoteCryptographyProvider to allow omitting the key version

## Why

A few reasons we're doing this: 
1. Discoverability - by surfacing this at the top level it'll make `CryptographyClient` more discoverable
2. Being able to share an httpClient and Pipeline instead of continuously creating new ones

For the second bullet-point, I noticed in passing that we validate that the version exists and is not empty. That's incorrect, we 
should allow version to be optional. The service supports it, all other languages do that, and our public API says version is
optional anyway.

Resolves #17529