Remove references to specific Node.js versions in readmes (#17703)
Now that we have a support policy around Node.js versions, issues were logged as part of a sweep of all the documentation to ensure we don't mentioned specific Node.js versions and instead point to the support policy. This PR does the needful for a handful of packages. Similar changes are needed for some arm packages that will be managed by the mgmt plane team. Another issue for Track 1 cognitive services is #17603 which will be handled separately

Fixes #17632
Fixes #17650
Fixes #17663
Fixes #17668
Fixes #17678