Update API Extractor for all packages (#17917)
## What

- Update API Extractor to the latest version (currently 7.18.11)
- Regenerate all API reviews by building all the packages

## Why

This is something we keep bumping into. First, we needed to upgrade API Extractor to allow us to re-export directly from
`@opentelemetry/api`. Then, it looks like we needed this to upgrade TypeScript to 4.4. 

We are way behind on this version, and it's time to upgrade.

## Callouts

How noisy is this?! Here's what's happening - somewhere around 7.9 I think API Extractor improved the way it detects name
collisions with predefined globals. Things like KeyType, Response, etc. 

If there's a clash, the generated API markdown file will rename <Item> to <Item_2> to make the name collision explicit.

Talking to folks on the team, and the poor souls that will be doing API Review approvals, we agreed that doing the upgrade
in one fell swoop is the way to go. 

Resolves #9410