[KeyVault] - Support multitenant challenge authentication (#18082)
## What

- Parse tenantId from WWW-Authenticate authorization field and pass it to the getToken call when parsing challenges

## Why

Identity will soon add multi-tenant challenge authentication support. It's already supported in `main` today, and will supported
in production soon. On the KeyVault side we just need to pull out the tenant ID from the `authorization` field and pass it in 
as part of the `GetTokenOptions` options bag when calling `getToken`. Older Identity clients will ignore this field, but newer
clients will use it as the tenantId.

Resolves #17619