[KeyVault] - Add missing lib:dom reference (#18344)
## What

- Add lib: dom entry to keyvault-secrets tsconfig
- Add lib: dom entry to keyvault-admin tsconfig
- Remove dom.d.ts from keyvault-common

## Why

During min/max testing we found that the dom.d.ts file doesn't actually help
when keyvault-common is bundled in with the rest of the source files. The other
keyvault packages are already including lib: dom which is why things worked fine
for them.

The solution here is to just add lib: dom to kv-secrets to match the other kv
packages. 

While not strictly necessary yet I did add it to keyvault-admin as well because
it'll surely come up once MHSM deployments resume correctly.