[core-http] HttpHeaders.clone() should keep raw header name casing  (#18348)
Original header names are still available in the values of
`_headersMap` and in the result of `headersArray()`.  However we
currently initialize the clone using `this.rawHeaders()` which returns
an object with lower-cased keys thus the original raw header names
are lost.

When cloning HttpHeaders, we should use the original header names in
the clone instead.