[KeyVault] - Fix sovereign cloud tests (#18350)
## What

- Suppress key rotation tests in sovereign clouds
- Ensure AZURE_AUTHORITY_HOST env var is passed to the credential in tests
- Suppress premium SKU tests when KV is not premium 

## Why

These are added to support sovereign cloud testing. 

The first is a must for us as key rotation is simply not available in all
clouds yet as a preview feature. We still want to test this feature in public cloud though so suppressing it in government
clouds is a reasonable approach.

The second is necessary to authenticate correctly in different authorities.

Finally, the third is a must as premium SKU is not available everywhere.

Fixes #18142 
Partially Fixes #18267 
Fixes #18189