[core-client] authorizeRequestOnClaimChallenge (#17315)
Not so long ago, Python exposed an `ARMChallengeAuthenticationPolicy ` and integrated it to their management core: https://github.com/Azure/autorest.python/pull/992 — this policy adds support for the CAE challenges on the management plane clients. While we haven’t heard of an exact date when this is going to be scoped for JS, we’ve been preparing for this moment for a while, so for us to stay up to date, it should be pretty simple.

This new `ARMChallengeAuthenticationPolicy` policy on Python’s source code can be seen here: https://github.com/Azure/azure-sdk-for-python/blob/9502e669e59e172a9d3d20586ba9761a64e23266/sdk/core/azure-mgmt-core/azure/mgmt/core/policies/_authentication.py — It essentially is the test code that we had in the `bearerTokenAuthenticationPolicy` tests that helped us verify the behavior of the challenge callbacks.

To solve for the same scenario, and after feedback on my initial implementations, I’m adding a new public method on the @azure/core-client called `authorizeRequestOnClaimChallenge`, which can be used as the `authorizeRequestOnChallenge` challenge callback on the `bearerTokenAuthenticationPolicy`. Once this is merged, the code generator for the ARM clients will need to change with something similar to:

```ts
import { bearerTokenAuthenticationPolicy } from "@azure/core-rest-pipeline";
import { authorizeRequestOnClaimChallenge } from "@azure/core-client";

const bearerTokenAuthenticationPolicy = bearerTokenAuthenticationPolicy({
  authorizeRequestOnChallenge: authorizeRequestOnClaimChallenge
});
```

If approved (eventually):

Fixes #17386

How does this look? Any feedback appreciated.