[KeyVault] - Update duration sample and doc comments (#18508)
## What

- Use Day.js in the Key Rotation sample
- Update the doc comments for Key Rotation

## Why

We discussed adding updating the API of Key Rotation to take an object instead
of a raw string for ISO8601 durations; however, with Temporal adding support for
Durations we want to avoid duplicating the type for now. Instead, we decided to
demonstrate interop with an existing 3rd party library. I chose Day.js for this
sample as it has a nice clean API.