[eslint] - fix plugin rule list (#18771)
It looks like ts-package-json-sdktype rule entry was duplicated in #18597 -
which causes the "number of rules should match the expected value" to report an
error.

This PR fixes the issue by removing the duplicated entry. I also went ahead and
sorted the entries to make these things easier to spot in the future (although
there is no guarantee sort will be preserved).