[core] Added logger parameters to bearerTokenAuthenticationPolicy (#18467)
Allowed customizing the internal logger of the `authorizeRequestOnClaimChallenge` challenge provided by the `@azure/core-client`. Client developers will need to send the logger through at the moment they initialize the `bearerTokenAuthenticationPolicy`. Currently, there are no other uses for this logger in the `bearerTokenAuthenticationPolicy`, but this opens us to the possibility of adding more logging in the future.

This is based on feedback from @xirzec here: https://github.com/Azure/azure-sdk-for-js/pull/17315#discussion_r740428119