Update using-azure-identity.md
Recently, we moved the authorization-code-credential sample into the manual tests folder, since this sample is really intended to test the credential and not to showcase a real world scenario (I plan on making improved samples with this credential in the upcoming months).

I didn’t notice that those changes were breaking a link we had at this level. Thanks to Daniel Jurek for letting me know about this issue.