[core] - Ensure package information is provided for traces (#18926)
## What

- Pass namespace and packageInformation as required options to createTracingClient
- Pass packageInformation as required options to `Instrumenter#startSpan`

## Why

When we discussed the API design for this, @lmolkova mentioned that ideally we'd require information about the package 
invoking the trace to ensure that data is passed along correctly to AZ Monitor. It was originally optional, and as part of the 
options bag for `createTracingClient` function.

But I accidentally made it "optionally required" on the Instrumenter level. Example:

```ts
instrumenter.startSpan("name") // OK
instrumenter.startSpan("name", {}) // Error: missing required packageInformation or something like that
```