[KeyVault] - Clarify certificate version properties (#18947)
## What

- Add comment about CertificateProperties#version which can only be populated by
the service
- Add explanation for updating the latest version of a certificate using an
empty version string

## Why

When updating some of our samples, Dina noticed a few things that can lead to
confusion. 

The first is that `version` is both a positional param _and_ part of
the `CertificateProperties` param. It's not clear which to use and its due to us
using the same model for both input and output types here.

The second is that version is technically required but not really - a user can
pass an empty string to update the latest version of a certificate. This is
common across all KV APIs but we did not make it truly optional.

Since these would be breaking changes, and no customers have yet to complain
about this, we decided a good middle-ground would be to update the documentation
to make it clearer.

Fixes #18886