[rollup] - Fix bug in circular dependency warning suppression (#19012)
## What

- Fix incorrect logic when suppressing chai's circular dependency warnings
- Move to the common dev-tool configuration where possible

## Why

This is a longstanding issue that we have, where an incorrect logic was copy-pasted to other places. I figured while cleaning this up that any package I touch can just convert over to the shared dev-tool configuration. Where I was unable
to do that, I just fixed this bug to avoid too many changes in one PR.

Fixes #14292
Resolves #17818 
Resolves #17816 
Resolves #17815 
Resolves #17814 
Resolves #17813 
Resolves #17810