[KeyVault] - Use shared rollup config (#19091)
## What

- Migrate keyvault to use the shared rollup config
- Teach the shared config to handle `<package>-common`  shared-source pattern
- Suppress `this has been rewritten to undefined` errors for OpenTelemetry
- Replace usages of `assert` with `chai`

## Why

We're migrating packages over to the shared config, and this handles Key Vault packages. 

Key Vault is the first migrated package that includes the shared-source keyvault-common pattern, and as such it does not conform to the file structure rollup config expects. So we
need to teach it to be a little smarter.

Finally, OTel's codebase emits the spreadArray helper ` var __spreadArray = (this && this.__spreadArray) || ...` which causes Rollup to be unhappy as the top-level `this` is 
undefined. In **this** case (couldn't resist) it's harmless and can be suppressed.

Finally, we are removing assert entirely so it doesn't make sense to add it to the namedExports collection - insteadI am just replacing it with `chai` wherever I encounter it.

Resolves #13410