[service-bus / eventhubs] - Fix browser tests (#19244)
## What

- Patch service-bus and event-hubs' rollup config with custom plugins that are required for test runs
- Replace `assert` with `chai`

## Why

We tried to use the shared rollup config across packages but unfortunately service-bus event-hubs are and will likely remain a special case due to the various dependencies it pulls in. We identified a few fixable issues but there are some
that will not be easily resolved.

So, working with Will we decided a reasonable compromise is to use the shared config but patch the plugins so that they
match what they were before the recent changes.

In passing, I am also removing any imports of the `assert` module as that has been removed and the tests are failing. So 
this should fix that as well.