[KeyVault] - Enable multi-version tests for KV secrets (#19354)
## What

- Add support for multi-version testing for keyvault-secrets

## Why

Our commitment is to support the last N versions of the KeyVault service. To ensure neither we nor the service introduce breaking changes to prior versions, multi-version test support was added. We ported keys over and have already reaped the benefits of version based filtering.

This commit adds the same support for keyvault-secrets

Resolves #17892