[apimanagement][test] reduce polling 202 responses in recordings
as they still cost a lot of time even the polling interval is set to 0 in playback. This PR removes most of the 202 responses while still keep some in order to test the polling.

when recording, it is probably better to set the polling interval to a large number instead of using the default.