[ServiceBus][Test] fix unstable tests that verify message state (#19357)
* [ServiceBus][Test] fix unstable tests that verify message state

The test messages we create have `scheduledEnqueueTimeUtc` set. Most of the time
messages received from the service have "active" state, but occasionally have
"scheduled" state due to the set property, thus failing the assertion.

This PR updates the check to check for either "active" or "schedule". In
addition, a test is added for "active" state where the `scheduledEnqueueTimeUtc`
property is removed before sending the message.

Also shown in the failing pipeline that when the test failed, the message wasn't
removed because `receiver.completeMessage()` is not called. The lingering
message caused the next test to fail. While updating the tests I also moved the
cleanup step before assertions so the message should always be removed.

* Undo changes to pipeline

* Address CR feedback

* Disable linting rule

* Remove an unstable assertion on message state.