[dev-tool] Handle passthrough of -- args (#19451)
Merely adding `--` args wasn't enough. Dev-tool's argument parsing is recursive, and it achieves this by passing the remainder of a branching-command's `args` back through the arg parser until a leaf-command is reached. Unfortunately, that means that if `--` args are parsed at the first step of the process, they aren't passed back into further subcommands' arguments. This meant that `--` had to be used by the first command. This change updates the argument passing behavior of `subCommand` to make it so that branching commands pass `--` arguments to their children.

I had previously tested to make sure that argument parsing worked as expected, but now I've also added a test to validate that `--` options are received by child commands.

@HarshaNalluru