Fix an issue where user error handler is not called
when subscribing to a valid topic but an invalid or disabled subscription.

The problem is that in these two cases, the service accepts our client's cbs
claim negotiation, and the receiver link is created. The [client receiver
enters](https://github.com/Azure/azure-sdk-for-js/blob/14099039a8009d6d9687daf65d22a998e3ad7920/sdk/servicebus/service-bus/src/core/streamingReceiver.ts#L539)
the `retryForeverFn` to call `_initAndAddCreditOperation()`, where the following

```ts
  await this._messageHandlers().postInitialize();
  this._receiverHelper.addCredit(this.maxConcurrentCalls);
```

will return successfully, despite that the fired-and-forgot `addCredit()` call
would later leads to `MessagingEntityDisabled` or `MessagingEntityNotFound`
errors in the underlying link. Since there's no errors thrown in our
retry-forever loop, the `onErrors()`
[callback](https://github.com/Azure/azure-sdk-for-js/blob/14099039a8009d6d9687daf65d22a998e3ad7920/sdk/servicebus/service-bus/src/core/streamingReceiver.ts#L541)
is not invoked. It is where the user error handler is called.

Because of the error, the link is detatched. We have code to re-initializing the
link when errors happen, so we call `_subscribeImpl()` where the
`retryForeverFn()` is called again. This goes on in an endless loop.

This PR adds code to invoke the user error handler in `_onAmqpError()` when the
error code is `MessagingEntityDisabled` or `MessagingEntityNotFound` so users
have a chance to stop the infinite loop.

There's another problem. When users call `close()` on the subscription in their
error handler, `this._receiverHelper.suspend()` is called to suspend the
receiver. However, when re-connecting we call `this._receiverHelper.resume()`
again in `_subscribeImpl()`. This essentially reset the receiver to be active
and we will not abort the attempt to initialize connection

https://github.com/Azure/azure-sdk-for-js/blob/14099039a8009d6d9687daf65d22a998e3ad7920/sdk/servicebus/service-bus/src/core/streamingReceiver.ts#L578

To fix it, this PR moves the `resume()` call out. It is supposed to only called
to enable receiver before subscribing. It should not be called when we try to
re-connect.