Fix serializer issue with nested polymorphics.
The Media Services SDK was having some serialization related issues because of some edge cases in how we handle polymorphic types.

Specifically, when a Sequence property of a polymorphic type has an element that is also a polymorphic type.

The issue stems from the fact that the `element` declaration inside Composite properties is a subset of the full definition, that is to say it *only* has the properties `name` and `className`, whereas the full Mapper would also declare things like `uberParent` and polymorphicDiscriminator`.

Because of this, we wouldn't correctly look up the polymorphic type of the elements in the array and instead serialize them as their parent type.

The subtlety is we *would* try to look up the discriminator, but without the `uberParent` definition, we would fall back to the current className, which would produce the wrong identifier for use in the `discriminators` map

Separately, I noticed we weren't properly deserializing items because of how `serializedName` gets escaped by the generator.

Furthermore, we again weren't correctly looking up the type, but this wasn't easily noticeable because the deserializer by default will preserve unknown properties, but those properties weren't getting correctly deserialized either.