Add generated folders to policheck exclusions (#19454)
## What

- Add `*generated*` folders to Policheck exclusions

## Why

Because generated code is generated directly from the swagger we should not flag
it as Policheck terms as that would be noisy, the upstream service swagger as
presumably already validated their terms, and really the service team owns the
terms here.

While we should continue to validate our codebase for global readiness, limiting
our checks to hand-authored and/or API review files only makes sense. Similar to
our recordings (which are also excluded).