[instrumentation] - Rename @azure/instrumentation-opentelemetry to @azure/opentelemetry-instrumentation-azure-sdk (#19629)
## What

- Rename the @azure/instrumentation-opentelemetry package to @azure/opentelemetry-instrumentation-azure-sdk
- Change spanContext from a getter to a function

## Why

The first change came from a review where we noticed that the package name is not entirely clear - is this azure instrumenting OTel? Are we instrumenting the Azure SDK or some other Azure service? 

Looking at other OTel instrumentation packages, it looks like this naming convention is pretty popular - see https://github.com/open-telemetry/opentelemetry-js-contrib/tree/main/plugins/node for reference

The second change came about when I was reviewing tracingPolicy - it looks like Span#spanContext is a function today, and there's no good reason to change that. In fact, it _might_ cause some issues in the future. So for consistency with existing APIs I am changing it to a function.