[eventgrid] fix distributed tracing enricher (#19625)
The cloud event distributed tracing enricher depends on `tracingPolicy` being
applied before it so it can copy the `traceparent` header that added by
`tracingPolicy` to its event payload. However, it doesn't specify this
dependency and relies on the implicit policy ordering when request pipeline is
created. This happened to work because `tracingPolicy` was not in a phase, and
was added before `cloudEventDistributedTracingEnricherPolicy` which is not in a
phase either. But after refactoring done in PR #19078, `tracingPolicy` is moved
to be after the Retry phase, which comes after policies that are not in a phase.
Now the `traceparent` has not been set when
`cloudEventDistributedTracingEnricherPolicy` is applied. This is not the
expected behavior and caused test failure.

This change fixes the issue by explicitly requiring the
`cloudEventDistributedTracingEnricherPolicy` to be applied after the
`tracingPolicy`.