[core-client] Handle calling onResponse when a PipelineRequest fails (#19702)
This change expands the functionality of `onResponse` to ensure that we don't forget to call it when a PipelineRequest fails, such as from an unexpected status code being detected in the deserialization policy. 

Though this is sometimes modeled as an exception (e.g. a service may decide to throw on 404 for a particular operation) we still want to provide the HTTP response to the caller. This change also widens the callback slightly to be able to take the error object explicitly. 

Fixes #18949