[EventGrid] Update System events and prepare for Jan release (#19741)
* [EventGrid] Event Schema Updates for January Release

This change pulls in the latest changes from the
`azure-rest-api-specs` repository for system events in Azure.

While no new events have been added with this release, some events now
have additional properties (the service teams which produce these
events now provide additional information in each event).

In addition, it fixes an issue we had with some of the Azure Resource
Manager events. We discovered that the event schemas for these events
was authored incorrectly and said some properties were of type
"string" instead of either a rich object or a map of strings to
strings.

A customer discovered this for .NET and that caused us to do some
investigation with the service team to discover that the specification
was wrong and get it corrected.

This is a breaking change for code which is written in typescript and
which uses any of these impacted properties because the underlying
type has changed from string to something else. However, any code
which previously was written against the old typings would have
*failed* at runtime with a type error, so in practice anyone who gets
an issue is now correctly discovering a bug in their program at
compile time now instead of runtime.

To prevent these issues in the future, we've been working closely with
the service team to introduce guard rails that ensure the events
services publish correspond to the schemas they have have published.

* [EventGrid] Prepare for Release