[core-rest-pipeline] Add Fetch HttpClient (#19530)
This PR adds a new HttpClient based on Fetch API.

For reporting progress I'm using 2 strategies depending on platform support. For platforms that support `TransformStream` we use those which are cleaner, however Firefox doesn't support `TransformStream` yet. In those cases we pump the incoming stream into a new stream in which we report progress.

This PR adds 3 main changes to the public surface

1. RequestBodyType can now also be a `ReadableStream<Uint8Array>`
2. `enableBrowserStreams` RequestOption which is off by default to keep current behavior in xhr client. When ON streamed responses will populate the new property in the PipelineResponse `browserStreamBody` instead of `blobBody`.
3.  new `browserStreamBody` property in PipelineResponse which contains the raw ReadableStream from the response body

As part of this change I'm setting the new fetch client as the default browser client. And will be running live tests for SDKs using core-rest-pipeline to validate functionality. 

**Note:** Identity browser tests needed to be updated as the mocks were tightly coupled with `xhr`, updated the test to mock `fetch` instead. Ideally, if possible, we should update these tests to be loosely coupled from the httpClient, however that is out of the scope of this PR

**Edit:** I was able to keep tests as unit-test without needing to bring in Autorest TestServer