[cosmos] Update the changelog for cosmos release
CC @sajeetharan, @jay-most : The release bot will take care of adding the new changelog entry for next release, and we want to make sure the current version is at the top of the changelog when the package is actually released. The change log is embedded into the package file that is posted to NPM, so it needs to be accurate and complete when it is published.

I also changed the release date to today.