[KeyVault] - Fix certificate poller test (#19909)
**Checklists** 
- [x] Added impacted package name to the issue description

**Issues associated with this PR:**
Fixes #18680

**Describe the problem that is addressed by this PR:**
Our china government cloud tests started failing intermittedly with a timeout.
Investigating further, I was able to reproduce this in China cloud but not in
Azure public cloud. I noticed that we aren't waiting for the certificate
creation poller in the test, leading to what is likely a race condition and
explains the intermittent failures.

After polling, I ran the same test multiple times and did not see any issues.