[core-xml] Fix root version node handling (#20035)
**Checklists** 
- [x] Added impacted package name to the issue description

**Packages impacted by this PR:**
@Azure/core-xml
@Azure/data-tables

**Issues associated with this PR:**
#20011 

**Describe the problem that is addressed by this PR:**
fast-xml-parser changed the way they handle the root <?xml node. In v3 they used to ignore it, now they add it to the model. This caused parsing to be incorrect impacting the tests in Tables SDK.

**What are the possible designs available to address the problem**
The fix I'm opting to is to detect the parsed "?xml" node and remove it from the parsed object before continuing.

**Are there test cases added in this PR?**_(If not, why?)_
Yes