[Identity] Moving @types/stoppable to the devDependencies (#20070)
`@types/stoppable` was added by mistake to the `dependencies` of the `package.json` of `@azure/identity`. This package is only used to build the TypeScript code, and not used on runtime, so it should have been placed on the `devDependencies` from the beginning. This package has no effect during runtime.

Fixes #20024