[dev-tool] Throw error when proxy fails to start (#19960)
Resolves #19725

When running tests using the test proxy, if the proxy tool image fails to start, the tool would exit successfully without outputting a useful error. This could happen if, for example, the Docker daemon is not running (this has happened to me a few times). This PR throws an error when the command to start the proxy tool outputs a nonzero exit code. This error will aid with debugging.