[KeyVault] - Added immutable flag to key release policy (#20090)
### Packages impacted by this PR
@azure/keyvault-keys

### Issues associated with this PR
Resolves #19857

### Describe the problem that is addressed by this PR
Now that this flag has been added we can regenerate code and add support for
immutable key release policies. Once a policy is marked immutable it can no
longer be modified.

### Are there test cases added in this PR? _(If not, why?)_
Yes, I debated whether I should also add a test case for "create a key, then
update the release policy to be immutable, then try to change it" but it
wouldn't drive any code so I skipped it.

Do note that since MHSM does not currently support this flag I was unable to add
a test for it for MHSM, but I will create an issue to add that in later.