[core-tracing] Fix the return type of withSpan (#20103)
### Packages impacted by this PR
@azure/core-tracing

### Issues associated with this PR
N/A - created after discussion with some team members

### Describe the problem that is addressed by this PR
`withSpan` is meant to await some user-provided callback and set the appropriate
status on the created span before closing it. Because we have no way of knowing
whether the callback is sync or async we must promisify it first. Up until now that
meant that the return type of a promise-returning callback was
`Promise<Promise<...>>` which does not reflect the reality that `await` will
recursively unwrap promises. That meant that the consumer must also either
`await` or mark the function as `async` in order to avoid a return-type
mismatch.

### What are the possible designs available to address the problem? If there are more than one possible design, why was the one in this PR chosen?
The solution here is to introduce an `Resolved` type which is a narrower
version of TypeScript 4.5's `Awaited` type. Because our min-bar is lower we
cannot take advantage of this new type yet, so I introduced something like it
which can be replaced with the broader type once we support TS >= 4.5.