[template] - Update @azure/template with new tracing APIs (#20120)
### Packages impacted by this PR
@azure/test-utils
@azure/template

### Issues associated with this PR
None

### Describe the problem that is addressed by this PR
As we start to onboard packages to the redesigned tracing APIs one of the best
places to start is the template project. This will allow me to observe whether
any builds break, or if anything goes awry, before other packages migrate over.

While adding test support, I decided to change the tracing test support to be a
bit more generic. When a chai plugin is consumed it often provides multiple
assertions, instead of a single assertion. In this way, users will not have to
import a new plugin every time we add custom assertions.

### What are the possible designs available to address the problem? If there are more than one possible design, why was the one in this PR chosen?
The only decision I had to make was whether to wait until
core-tracing@preview.14 is published on NPM. I decided to open this PR now so I
can see if there's any impact of depending on an unpublished version.