[CODEOWNERS] - Make primary package owner first (#20155)
### Packages impacted by this PR
None

### Issues associated with this PR
Resolves #18210

### Describe the problem that is addressed by this PR
Today, vendors create issues and assign them based on a list that I provided
them. The problem is that it can easily get out of sync as nobody would know to
update it. I am working towards making CODEOWNERS the single source of truth for
package ownership. It needs to start with the primary owner listed first, so
that the vendors know who to assign it to.

### What are the possible designs available to address the problem? If there are more than one possible design, why was the one in this PR chosen?
I am going to be honest, this effort might not go anywhere. But it's not a
harmful thing to do. Besides, there's a bit of a chicken-and-egg problem right
now. CODEOWNER list ordering is not used by the vendors because it's not
ordered, but nobody will want to order it because the order has no impact.
Between this and the efforts engsys is putting in place to parse this file I am
hopeful we can get to a place where vendors can look at data derived from
CODEOWNERS.