[EventGrid] Update System events and prepare for Feb release
This change pulls in the latest changes from the
`azure-rest-api-specs` repository for system events in Azure.

It fixes two issues with our typings for events (a property name was
incorrectly cased and another property was typed as a boolean when the
event delivered by the service uses a string (with the value of `true`
or `false`) instead of a boolean as might be expected.

This is a breaking change for code which is written in typescript and
which uses any of these impacted properties because the underlying
type has changed from string to something else. However, any code
which previously was written against the old typings would have
*failed* at runtime with a type error, so in practice anyone who gets
an issue is now correctly discovering a bug in their program at
compile time now instead of runtime.