[cosmos] Remove dependency on snap-shot-it (#20259)
Fixes #19306

This is the only place we use the `snap-shot-it` dependency. It has caused trouble by not keeping its own dependencies up to date in the past, so this PR removes it entirely.

For simplicity, and to separate the tests from the snapshot, I moved the snapshot to be near the test (with some small modifications) and added a helper which replaces the functionality of `snap-shot-it`.